requirements.txt
arrow==0.11.0
cachetools==3.1.1
certifi==2019.6.16
chardet==3.0.4
Click==7.0
Flask==1.1.1
Flask-WTF==0.14.2
google-api-python-client==1.7.11
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.0
httplib2==0.13.1
ics==0.5
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
oauthlib==3.1.0
pyasn1==0.4.6
pyasn1-modules==0.2.6
python-dateutil==2.8.0
python-dotenv==0.10.3
requests==2.22.0
requests-oauthlib==1.2.0
rsa==4.0
six==1.12.0
uritemplate==3.0.0
urllib3==1.25.3
Werkzeug==0.15.5
WTForms==2.2.1
